feat: merge EditorBlock and useComponent
